Add AbstractTokenProvider
