+++ src/DependencyInjection/Configuration.php
@@ -23,7 +23,7 @@ class Configuration implements ConfigurationInterface
                 ->booleanNode('enabled')->defaultTrue()->end()
                 ->scalarNode('field_name')->defaultValue('_token')->end()
                 ->scalarNode('token_provider_class')->defaultValue('Genedys\\CsrfRouteBundle\\Routing\\TokenProvider\\TokenProvider')->end()
-                ->scalarNode('token_provider_base_class')->defaultValue('Genedys\\CsrfRouteBundle\\Routing\\TokenProvider\\TokenProvider')->end()
+                ->scalarNode('token_provider_base_class')->defaultValue('Genedys\\CsrfRouteBundle\\Routing\\TokenProvider\\AbstractTokenProvider')->end()
                 ->scalarNode('token_provider_dumper_class')->defaultValue('Genedys\\CsrfRouteBundle\\Routing\\TokenProvider\\Dumper\\TokenProviderDumper')->end()
                 ->scalarNode('token_provider_cache_class')->defaultValue('%router.cache_class_prefix%CsrfTokenProvider')->end()
             ->end()
