+++ src/Routing/TokenProvider/TokenProvider.php
@@ -3,68 +3,45 @@
 namespace Genedys\CsrfRouteBundle\Routing\TokenProvider;
 
 use Genedys\CsrfRouteBundle\Model\CsrfToken;
-use Genedys\CsrfRouteBundle\Routing\TokenProviderInterface;
 use Symfony\Component\Routing\Route;
 use Symfony\Component\Routing\RouteCollection;
 
 /**
  * @author Jáchym Toušek <enumag@gmail.com>
  */
-abstract class TokenProvider implements TokenProviderInterface
+final class TokenProvider extends AbstractTokenProvider
 {
     /**
-     * @var string
+     * @var RouteCollection
      */
-    private $fieldName;
+    private $routeCollection;
 
     /**
      * @param string $fieldName
      */
-    public function __construct($fieldName)
+    public function __construct($fieldName, RouteCollection $routeCollection)
     {
-        $this->fieldName = $fieldName;
+        parent::__construct($fieldName);
+
+        $this->routeCollection = $routeCollection;
     }
 
     /**
-     * @return CsrfToken
-     */
-    protected function getDefaultToken()
-    {
-        $token = new CsrfToken();
-        $token->setToken($this->fieldName);
-        $token->setIntention(null);
-        $token->setMethods('GET');
-
-        return $token;
-    }
-
-    /**
-     * @param true|array $option
+     * Returns CSRF configuration for the given route.
+     *
+     * @param string $name
      * @return CsrfToken|null
      */
-    protected function getTokenFromOption($option)
+    public function getCsrfToken($name)
     {
-        if (true === $option) {
-            return $this->getDefaultToken();
-        }
-
-        if (!is_array($option)) {
-            return null;
-        }
-
-        $token = new CsrfToken();
-        $token->setToken(array_key_exists('token', $option) ? $option['token'] : $this->fieldName);
-        $token->setIntention(array_key_exists('intention', $option) ? $option['intention'] : null);
-        $token->setMethods(array_key_exists('methods', $option) ? $option['methods'] : 'GET');
-
-        return $token;
+        $this->getTokenFromRoute($this->routeCollection->get($name));
     }
 
     /**
      * @param Route $route
      * @return CsrfToken|null
      */
-    protected function getTokenFromRoute(Route $route)
+    private function getTokenFromRoute(Route $route)
     {
         // Check if route has the option
         if (!$route->hasOption(self::OPTION_NAME)) {
